Move `QueryActionEditor` to `QueryActionContext`
