+++ frontend/src/metabase/actions/containers/ActionCreator/ActionContext/QueryActionContextProvider/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./QueryActionContextProvider";
 export * from "./QueryActionContextProvider";
+export { ACE_ELEMENT_ID } from "./QueryActionEditor";
